FRESH_START/requirements.txt
pypdf
pdfplumber
tqdm
google-genai
google-generativeai
google-cloud-storage
google-cloud-aiplatform
google-auth
vertex-ai
langchain
langchain-core
langchain-google-genai
langchain-community
chromadb
fastapi
uvicorn
pydantic
